+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Make FOLIO number read only. Fixes UIU-348.
 * Use consistent language for "no `<attribute>` found" messages. Fixes UX-115.
 * Hide proxy section if user lacks sufficient permission to see it. Fixes UIU-363.
+* Add Save button to user settings. Fixes UIU-354.
 
 ## [2.12.0](https://github.com/folio-org/ui-users/tree/v2.12.0) (2017-11-28)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.11.0...v2.12.0)